requirements.txt
Flask==1.1.4
pillow==10.2.0
requests==2.31.0

blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==7.1.2
colorama==0.4.6
idna==3.6
importlib-metadata==7.0.1
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==2.0.1
urllib3==2.2.0
Werkzeug==1.0.1
zipp==3.17.0
